+++ src/components/common/Header.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import * as S from "../../App.styles";
+import ServerTimeComponent from "../ServerTime";
+import { ReservationListProps } from "../../declare/type";
+
+export const Header = ({ reservationList }: ReservationListProps) => {
+  return (
+    <S.TitleWrapper>
+      <S.TitleLogoImage src={"logo.jpg"} />
+      <ServerTimeComponent />
+    </S.TitleWrapper>
+  );
+};
